LPD-85191 Move utils to object-web
